refactor(dongtai-common): modify remote config entity and collect to PerformanceLimitThreshold;
